third_party/blink/web_tests/fast/url/relative-unix-expected.txt

Test resolution of relative UNIX-like URLs.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS canonicalize('\\\\Another\\path') is 'http://another/path'
PASS canonicalize('/c:\\foo') is 'http://host/c:/foo'
PASS canonicalize('//c:\\foo') is 'http://c/foo'
PASS successfullyParsed is true

TEST COMPLETE
